requirements.txt
asgiref==3.4.1
backcall==0.2.0
decorator==5.1.0
Django==3.2.9
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-mptt==0.13.4
django-polymorphic==3.0.0
ipython==7.29.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
pytz==2021.3
sqlparse==0.4.2
traitlets==5.1.1
wcwidth==0.2.5
